Can now detect when a file hasn't been fully downloaded. Also makes sure echocenter is loaded before doing stuff with it.
We use urlopen to check for the size of the file on the server and compare it against the local file size. If the web server size is bigger (or wasn't there), we download the file again. This catches where the download might have crashed or been cancelled by the user.

The selenium .get() doesn't detect whether the echocenter is loaded (it's in a frame or something) so we now catch that too in a try loop.