Fixed non-visible element error when list longer than window.
When the list of lectures grew longer than the space for it, the script would fail to select elements not in the visible space. I now catch this error and scroll the window down.
When trying to get the link for the download, if the element isn't yet loaded I catch the error and wait for a little bit. This allowed me to remove the conservative sleep(2) and sleep(1) calls before and after this block, speeding up gathering of download links significantly.